検索に関するシングルアクションコントローラーをグルーピング/#124
